enhancement:layout_car:use transparency to draw some map features (#879)
With the introduction of the multipolygon code we now have all the nice landuses in the map. However, in OSM there are some map features that are not mapped as multipolygon but just "above" the other landuses. Like poly_zoo, poly_theme_park or poly_airfield.
Current configuration causes them to be hidden by the landuses, as these are drawn after the above mentioned. OSM's Mapnik style solves the problem by drawing those feature transparent above the landuses.
This pull request does so for some map features as well causing the map to look way better.

If transparent drawing is not supported by the platform, this causes the mentioned polys to be drawn above the landuses, showing them, eventually hiding some landuse details underneath. But this looks even better than before too.

NOTE: Transparent drawing is known to work on: Qt5 (sailfish) and gtk.
NOTE: Transparent drawing is known NOT to work on SDL
Dont't know for all the others. Sombody might want to check Android? Remember #852 on comparing screenshots though.

* Fix: layout_car: draw map features transparent on top

Some map featuresa re usually not put into multipolygons, but simple
tagged "on top" of the others. So It's quite common for example to
tag "poly_airport" on top of the underlying other landuses inside the
airport. Move (some) of those elements on top of the polygon draw stack
and make them transparent. Looks way better on graphics that support
transparency. Not much is lost on those that don't.

* Fix: layout_car: differentiate meadow and farmland in color

* enhancement:layout-car:Make poly_barracks and poly_university transparent