fix:maptool:better map aerodroms and military areas (#1022)
* fix:maptool:better map aerodroms and military areas

This commit uses more complex rule set to map aerodromes and military
landuses depending on tag combination avoiding duplication of the area
in binfile.

For example military airports are quite often tagged:
landuse=military
military=airfield
aeroway=aerodrome
which caused the airport to be tripled in map. This is now resolved.

* Enhancement:layout_car:add stripesto military areas

* Fix:maptool:treat archaeological_site limes special

archaeological_site tag is used on way despite outruled by OSM wiki on
limes archaeological site. But there is a special tagging scheme for
Limes. Use that to fix rendering of Limes in map.